Restructuring of effects
This splits the effect stuff into seperate files for easier navigation.
